+++ NEWS.md
@@ -2,4 +2,6 @@
 
 ## 0.1.0
 
+* Added main canvas page screenshot saving to PNG
 * Added a `NEWS.md` file to track changes to the package.
+